+++ server/index.js
@@ -16,24 +16,14 @@ const http = require('http');
 const express = require('express');
 const consola = require('consola');
 const {Nuxt, Builder} = require('nuxt');
-const MongoClient = require('mongodb').MongoClient;
 const universalAnalytics = require('universal-analytics');
-const rp = require(`request-promise`);
-const {logger, getUrlBaseByWiki} = require('./common');
-const {root} = require('./routes/root');
-const {diff, diffWikiRevId} = require('./routes/diff');
-
+const rp = require('request-promise');
+const {db, logger, apiLogger, getUrlBaseByWiki, computeOresField, fetchRevisions, useOauth} = require('./common');
+const routes = require('./routes');
 const asyncHandler = fn => (req, res, next) =>
     Promise
         .resolve(fn(req, res, next))
         .catch(next);
-
-
-const apiLogger = new (require('heroku-logger').Logger)({
-  level: process.env.LOG_LEVEL || 'debug',
-  delimiter: " | ",
-  prefix: 'api'
-});
 
 const perfLogger = new (require('heroku-logger').Logger)({
   level: process.env.LOG_LEVEL || 'debug',
@@ -65,366 +55,11 @@ const logReqPerf = function (req, res, next) {
 
 let docCounter = 0;
 let allDocCounter = 0;
-function isEmpty(value) {
-  return typeof value == 'string' && !value.trim() || typeof value == 'undefined' || value === null;
-}
 
-const useStiki = !isEmpty(process.env.STIKI_MYSQL);
-const useOauth = !isEmpty(process.env.MEDIAWIKI_CONSUMER_SECRET) && !isEmpty(process.env.MEDIAWIKI_CONSUMER_KEY);
 console.log(`XXX process.env.MEDIAWIKI_CONSUMER_SECRET = ${process.env.MEDIAWIKI_CONSUMER_SECRET}, process.env.MEDIAWIKI_CONSUMER_KEY = ${process.env.MEDIAWIKI_CONSUMER_KEY}`)
 // Import and Set Nuxt.js options
 const config = require('../nuxt.config.js');
 config.dev = !(process.env.NODE_ENV === 'production');
-
-function computeOresField(oresJson, wiki, revId) {
-  let damagingScore = oresJson.damagingScore || (oresJson[wiki].scores[revId].damaging.score && oresJson[wiki].scores[revId].damaging.score.probability.true);
-  let badfaithScore = oresJson.badfaithScore || (oresJson[wiki].scores[revId].goodfaith.score && oresJson[wiki].scores[revId].goodfaith.score.probability.false);
-  let damaging = oresJson.damaging || (oresJson[wiki].scores[revId].damaging.score && oresJson[wiki].scores[revId].damaging.score.prediction);
-  let badfaith = oresJson.badfaith || (oresJson[wiki].scores[revId].goodfaith.score && !oresJson[wiki].scores[revId].goodfaith.score.prediction);
-  return {
-    wikiRevId: `${wiki}:${revId}`,
-    damagingScore: damagingScore,
-    damaging: damaging,
-    badfaithScore: badfaithScore,
-    badfaith: badfaith
-  }
-}
-
-async function fetchOres(wiki, revIds) {
-  let oresUrl = `https://ores.wmflabs.org/v3/scores/${wiki}/?models=damaging|goodfaith&revids=${revIds.join('|')}`;
-  let oresResultJson = await rp.get(oresUrl, {json: true});
-  return revIds.map(revId => computeOresField(oresResultJson, wiki, revId));
-}
-
-async function fetchRevisions(wiki, revIds = []) {
-  if (revIds.length > 0) {
-    const fetchUrl = new URL(`${getUrlBaseByWiki(wiki)}/w/api.php`);
-    let params = {
-      "action": "query",
-      "format": "json",
-      "prop": "revisions|info",
-      "indexpageids": 1,
-      "revids": revIds.join('|'),
-      "rvprop": "ids|timestamp|flags|user|tags|size|comment",
-      "rvslots": "main"
-    };
-    Object.keys(params).forEach(key => {
-      fetchUrl.searchParams.set(key, params[key]);
-    });
-    let retJson = await rp.get(fetchUrl, {json: true});
-    if (retJson.query.badrevids) {
-      return []; // does not find
-    }
-    else {
-      /** Example
-      {
-          "batchcomplete": "",
-          "query": {
-              "pageids": [
-                  "16377",
-                  "103072"
-              ],
-              "pages": {
-                  "16377": {
-                      "pageid": 16377,
-                      "ns": 104,
-                      "title": "API:Query",
-                      "revisions": [
-                          {
-                              "revid": 3319487,
-                              "parentid": 3319442,
-                              "minor": "",
-                              "user": "Shirayuki",
-                              "timestamp": "2019-07-16T22:08:58Z",
-                              "size": 15845,
-                              "comment": "",
-                              "tags": []
-                          }
-                      ],
-                      "contentmodel": "wikitext",
-                      "pagelanguage": "en",
-                      "pagelanguagehtmlcode": "en",
-                      "pagelanguagedir": "ltr",
-                      "touched": "2019-07-17T03:10:17Z",
-                      "lastrevid": 3319487,
-                      "length": 15845
-                  },
-                  "103072": {
-                      "pageid": 103072,
-                      "ns": 104,
-                      "title": "API:Lists/All",
-                      "revisions": [
-                          {
-                              "revid": 2287599,
-                              "parentid": 2287488,
-                              "user": "Wargo",
-                              "timestamp": "2016-11-17T16:49:59Z",
-                              "size": 924,
-                              "comment": "Undo revision 2287488 by [[Special:Contributions/Jkmartindale|Jkmartindale]] ([[User talk:Jkmartindale|talk]])",
-                              "tags": []
-                          }
-                      ],
-                      "contentmodel": "wikitext",
-                      "pagelanguage": "en",
-                      "pagelanguagehtmlcode": "en",
-                      "pagelanguagedir": "ltr",
-                      "touched": "2019-07-18T04:14:03Z",
-                      "lastrevid": 2287599,
-                      "length": 924
-                  }
-              }
-          }
-      }
-       */
-      let revIdToRevision = {};
-      for (let pageId of retJson.query.pageids) {
-        for (let revision of retJson.query.pages[pageId].revisions) {
-          revIdToRevision[revision.revid] = revision;
-          revIdToRevision[revision.revid].title = retJson.query.pages[pageId].title;
-          revIdToRevision[revision.revid].wiki = wiki;
-          revIdToRevision[revision.revid].wikiRevId = `${wiki}:${revision.revid}`;
-          revIdToRevision[revision.revid].pageLatestRevId = retJson.query.pages[pageId].lastrevid;
-          revIdToRevision[revision.revid].namespace = revision.ns;
-        }
-      }
-      return revIds.map(revId => revIdToRevision[revId]);
-    }
-  }
-}
-
-async function getNewJudgementCounts(db, matcher = {}, offset = 0, limit = 10) {
-  return await db.collection(`Interaction`).aggregate([
-        {
-          $match: matcher
-        },
-        {
-          "$group" : {
-            "_id" : {
-              "wikiRevId" : "$wikiRevId"
-            },
-            "wikiRevId" : {
-              "$first" : "$wikiRevId"
-            },
-            "judgements" : {
-              "$push" : {
-                "judgement" : "$judgement",
-                "userGaId" : "$userGaId",
-                "timestamp" : "$timestamp"
-              }
-            },
-            "totalCounts" : {
-              "$sum" : 1
-            },
-            "shouldRevertCounts" : {
-              "$sum" : {
-                "$cond" : [
-                  {
-                    "$eq" : [
-                      "$judgement",
-                      "ShouldRevert"
-                    ]
-                  },
-                  1,
-                  0
-                ]
-              }
-            },
-            "notSureCounts" : {
-              "$sum" : {
-                "$cond" : [
-                  {
-                    "$eq" : [
-                      "$judgement",
-                      "NotSure"
-                    ]
-                  },
-                  1,
-                  0
-                ]
-              }
-            },
-            "looksGoodCounts" : {
-              "$sum" : {
-                "$cond" : [
-                  {
-                    "$eq" : [
-                      "$judgement",
-                      "LooksGood"
-                    ]
-                  },
-                  1,
-                  0
-                ]
-              }
-            },
-            "lastTimestamp" : {
-              "$max" : "$timestamp"
-            },
-            "recentChange": {
-              "$first": "$recentChange"
-            }
-          }
-        },
-        {
-          "$project" : {
-            "wikiRevId" : "$_id.wikiRevId",
-            "judgements" : "$judgements",
-            "recentChange" : 1,
-            "lastTimestamp" : 1,
-            "counts.Total" : "$totalCounts",
-            "counts.ShouldRevert" : "$shouldRevertCounts",
-            "counts.NotSure" : "$notSureCounts",
-            "counts.LooksGood" : "$looksGoodCounts"
-          }
-        },
-        {
-          "$match" : {
-            "recentChange.ores" : {
-              "$exists" : true,
-              "$ne" : null
-            },
-            "recentChange.wiki" : {
-              "$exists" : true,
-              "$ne" : null
-            },
-            "lastTimestamp" : {
-              "$exists" : true,
-              "$ne" : null
-            }
-          }
-        },
-        {
-          "$sort" : {
-            "lastTimestamp" : -1
-          }
-        },
-      ],
-      {
-        "allowDiskUse": true
-      })
-      .skip(offset)
-      .limit(limit)
-      .toArray();
-
-  /**
-   * Example of output schema:
-   {
-      "_id":{
-        "wikiRevId":"enwiki:905704873"
-      },
-
-      "lastTimestamp":"1562791937",
-      "recentChange":{
-        "_id":"enwiki-1169325920",
-        "id":"1169325920",
-        "title":"Financial endowment",
-        "namespace":"0",
-        "revision":{
-          "new":"905704873",
-          "old":"900747399"
-        },
-        "ores":{
-          "damagingScore":"0.8937388482947232",
-          "damaging":"true",
-          "badfaithScore":"0.8787846798198944",
-          "badfaith":"true"
-        },
-        "user":"<userId or IP>>",
-        "wiki":"enwiki",
-        "timestamp":"1562791912"
-      },
-      "wikiRevId":"enwiki:905704873",
-      "judgements":[
-        {
-          "judgement":"ShouldRevert",
-          "userGaId":"<userGaId>",
-          "timestamp":"1562791937"
-        }
-      ],
-      "counts":{
-        "Total":1,
-        "ShouldRevert":1,
-        "NotSure":0,
-        "LooksGood":0
-      }
-   }
-   */
-}
-
-/**
- * @deprecated use getNewJudgementCounts
- * @param db
- * @param newRecentChange
- * @returns {Promise<void>}
- */
-async function getJudgementCounts(db, newRecentChange) {
-  let judgementCounts = {};
-  let aggrRet = await db.collection(`Interaction`).aggregate([
-        {
-          "$match": {
-            "recentChange.id": newRecentChange.id.toString(),
-            "recentChange.wiki": newRecentChange.wiki
-          }
-        },
-        {
-          // Counts group by Judgement
-          "$group": {
-            "_id": "$judgement",
-            "judgement": {
-              "$sum": 1.0
-            }
-          }
-        }
-      ],
-      {
-        "allowDiskUse": false
-      }).toArray();
-  aggrRet.forEach(ret => {
-    judgementCounts[ret._id] = ret.judgement;
-  });
-  return judgementCounts;
-}
-
-/**
- * @deprecated use queryMarkedRevisions
- * @param db
- * @param myGaId
- * @returns {Promise<any[]>}
- */
-async function queryMarkedRecentChange(db, myGaId) {
-  let recentChanges;
-  let interactions = await db.collection(`Interaction`).find({}, {
-    sort: [["timestamp", -1]]
-  })
-  // .limit(1)  // TODO add limit when performance becomes a problem, a typical case that RDB is better than NonRDB
-      .toArray();
-
-  let dbIds = interactions.map(rc => `${rc.recentChange.wiki}-${rc.recentChange.id}`);
-  recentChanges = await db.collection(`MediaWikiRecentChange`)
-      .find(
-          {"_id": {$in: dbIds}})
-      // .limit(1)  // TODO add limit when performance becomes a problem, a typical case that RDB is better than NonRDB
-      .toArray();
-
-  recentChanges = await Promise.all(recentChanges.map(async (rc) => {
-    // TODO improve performance.
-    rc.judgementCounts = await getJudgementCounts(db, rc);
-    return rc;
-  }));
-
-  // Add my judgement
-  if (myGaId) {
-    let myInteractionMap = {};
-    let myInteractions = await db.collection(`Interaction`).find({userGaId: myGaId}).toArray();
-    myInteractions.forEach(interaction => {
-      myInteractionMap[interaction.recentChange._id] = interaction;
-    });
-    recentChanges.forEach(rc => {
-      if (myInteractionMap[rc._id]) rc.judgement = myInteractionMap[rc._id].judgement;
-    });
-  }
-  return recentChanges.reverse();
-}
 
 // -------------- FROM STIKI API -------------
 function setupSTikiApiLisenter(app) {
@@ -501,279 +136,48 @@ function setupApiRequestListener(db, io, app) {
 
   apiRouter.use(cache('1 week', onlyGet));
 
-  apiRouter.get('/', root);
+  apiRouter.get('/', routes.root);
 
-  apiRouter.get('/diff/:wikiRevId', asyncHandler(diffWikiRevId));
+  apiRouter.get('/diff/:wikiRevId', asyncHandler(routes.diffWikiRevId));
 
-  apiRouter.get('/diff', asyncHandler(diff));
+  apiRouter.get('/diff', asyncHandler(routes.diff));
 
-  apiRouter.get('/recentChanges', asyncHandler(async (req, res) => {
-    logger.debug(`req.query`, req.query);
+  apiRouter.get('/recentChanges', asyncHandler(routes.recentChanges));
 
-    // let diffApiUrl = `${req.query.serverUrl}/w/api.php?action=compare&fromrev=${req.query.revId}&torelative=prev&format=json`;
-    // let diffJson = await rp.get(diffApiUrl, {json: true});
-    res.send(req.query);
+  apiRouter.get('/ores', asyncHandler(routes.ores));
 
-  }));
+  apiRouter.get('/ores/:wikiRevId', asyncHandler(routes.oresWikiRevId));
 
-  apiRouter.get('/ores', asyncHandler(async (req, res) => {
-    let revIds = req.query.revIds;
-    let wiki = req.query.wiki;
-    let ret = await fetchOres(wiki, revIds);
-    res.send(ret);
-    req.visitor
-        .event({ec: "api", ea: "/ores"})
-        .send();
-  }));
+  apiRouter.get('/revision/:wikiRevId', asyncHandler(routes.revisionWikiRevId));
 
-  apiRouter.get('/ores/:wikiRevId', asyncHandler(async (req, res) => {
-    let wikiRevId = req.params.wikiRevId;
-    let wiki = wikiRevId.split(':')[0];
-    let revId = wikiRevId.split(':')[1];
-    let ret = await fetchOres(wiki, [revId]);
-    if (ret.length === 1) {
-      res.send(ret[0] );
-    } else if (ret.length === 0) {
-      res.status(404);
-      res.send(`Can't find ores`);
-    } else {
-      res.status(500);
-      res.send(`Something is wrong`);
-    }
-    req.visitor
-        .event({ec: "api", ea: "/ores/:wikiRevId"})
-        .send();
-  }));
+  apiRouter.get('/revisions', asyncHandler(routes.revision));
 
-  apiRouter.get('/revision/:wikiRevId', asyncHandler(async (req, res) => {
-    let wikiRevId = req.params.wikiRevId;
-    let revisions = await fetchRevisions(wikiRevId.split(':')[0], [wikiRevId.split(':')[1]]);
-    if (revisions.length === 1) {
-      res.send(revisions[0] );
-    } else if (revisions.length === 0) {
-      res.status(404);
-      res.send(`Can't find revisions`);
-    } else {
-      res.status(500);
-      res.send(`Something is wrong`);
-    }
+  apiRouter.get('/interaction/:wikiRevId', asyncHandler(routes.getInteraction));
 
-    req.visitor
-        .event({ec: "api", ea: "/revision/:wikiRevId"})
-        .send();
-  }));
+  apiRouter.get('/interactions', asyncHandler(routes.listInteractions));
 
-  apiRouter.get('/revisions', asyncHandler(async (req, res) => {
-    let revIds = req.query.revIds;
-    let wiki = req.query.wiki;
-    let revisions = await fetchRevisions(wiki, revIds);
-    res.send(revisions);
-
-    req.visitor
-        .event({ec: "api", ea: "/revision/:wikiRevId"})
-        .send();
-  }));
-
-  apiRouter.get('/interaction/:wikiRevId', asyncHandler(async (req, res) => {
-    let wikiRevId = req.params.wikiRevId;
-    let interactions = await getNewJudgementCounts(db,
-    // {
-    //   wikiRevId: {$in: [wikiRevId]}
-    // },
-        { "wikiRevId": {"$in":[wikiRevId]} }
-        );
-    if (interactions.length >= 1) {
-      res.send(interactions[0] );
-    } else {
-      res.send({
-        "_id": {
-          "wikiRevId": wikiRevId
-        },
-        "wikiRevId": wikiRevId,
-        "judgements": [
-        ],
-        "counts": {
-          "Total": 0,
-          "ShouldRevert": 0,
-          "NotSure": 0,
-          "LooksGood": 0
-        }
-      });
-    }
-    req.visitor
-        .event({ec: "api", ea: "/interaction/:wikiRevId"})
-        .send();
-  }));
-
-  apiRouter.get('/interactions', asyncHandler(async (req, res) => {
-    let limit = parseInt(req.query.limit) || 10;
-    let offset = parseInt(req.query.offset) || 0;
-    let matcher = {};
-    if (req.query.wikiRevIds) {
-      matcher.wikiRevId = { $in: req.query.wikiRevIds }
-    }
-    if (req.query.userGaIds) {
-      matcher.userGaId = { $in: req.query.userGaIds }
-    }
-
-    let interactions = await getNewJudgementCounts(db, matcher, offset, limit);
-    res.send(interactions);
-    req.visitor
-        .event({ec: "api", ea: "/interactions"})
-        .send();
-  }));
-
-  apiRouter.post('/interaction/:wikiRevId', asyncHandler(async (req, res) => {
-    let userGaId = req.body.gaId;
-    let wikiRevId = req.params.wikiRevId;
-    /**
-    {
-      gaId: gaId,
-      judgement: this.myJudgement,
-      timestamp: Math.floor(new Date().getTime() / 1000),
-      wikiRevId: revision.wikiRevId,
-      recentChange: {
-        title: revision.title,
-        namespace: revision.namespace,
-        revision: {
-          new: revision.revid,
-          old: revision.parentid,
-        },
-        ores: this.ores,
-        user: revision.user,
-        wiki: revision.wiki,
-        timestamp: revision.timestamp
-      }
-    }
-    */
-
-    let doc = req.body;
-    await db.collection(`Interaction`).findOneAndReplace({
-      userGaId: userGaId,
-      wikiRevId: wikiRevId,
-    }, doc, {upsert: true});
-    apicache.clear(req.originalUrl);
-    let storedInteractions = await getNewJudgementCounts(db, {wikiRevId: {$in : [wikiRevId]}});
-    let storedInteraction = storedInteractions[0];
-
-    storedInteraction.newJudgement = {
-      userGaId:  doc.userGaId,
-      judgement: doc.judgement,
-      timestamp: doc.timestamp
-    };
-    io.sockets.emit('interaction', storedInteraction);
-
-    res.send(`ok`);
-    req.visitor
-        .event({ec: "api", ea: "/interaction"})
-        .event("judgement", req.body.judgement)
-        .send();
-  }));
+  apiRouter.post('/interaction/:wikiRevId', asyncHandler(routes.updateInteraction));
 
   /**
    * TODO fix it.
    * @deprecated
    */
-  apiRouter.post('/interaction', asyncHandler(async (req, res) => {
-    logger.debug(`Interaction req`, req.cookies, req.body);
+  apiRouter.post('/interaction', asyncHandler(routes.interaction));
 
-    let userGaId = req.body.gaId;
-    let newRecentChange = req.body.newRecentChange;
-    let doc = {
-      userGaId: userGaId,
-      wikiRevId: `${newRecentChange.wiki}:${newRecentChange.revision.new}`,
-      judgement: req.body.judgement,
-      timestamp: req.body.timestamp,
-      recentChange: {
-        _id: newRecentChange._id,
-        id: newRecentChange.id,
-        title: newRecentChange.title,
-        namespace: newRecentChange.namespace,
-        revision: newRecentChange.revision,
-        ores: newRecentChange.ores,
-        user: newRecentChange.user,
-        wiki: newRecentChange.wiki,
-        timestamp: newRecentChange.timestamp,
-      }
-    };
-
-    await db.collection(`Interaction`).findOneAndReplace({
-      userGaId: userGaId,
-      "recentChange.id": newRecentChange.id,
-      "recentChange.wiki": newRecentChange.wiki
-    }, doc, {upsert: true});
-
-    doc.judgementCounts = await getJudgementCounts(db, newRecentChange);
-    io.sockets.emit('interaction', doc);
-
-    res.send(`ok`);
-    req.visitor
-        .event({ec: "api", ea: "/interaction"})
-        .event("judgement", req.body.judgement)
-        .send();
-  }));
-
-  apiRouter.get("/markedRevs.csv", asyncHandler(async (req, res) => {
-    let newJudgementCounts = await getNewJudgementCounts(
-        db, {}, 0, 10000000/* as many as possible to download all */);
-    res.setHeader('Content-Type', 'text/csv');
-    res.setHeader('Content-Disposition', 'attachment; filename=\"' + 'download-' + Date.now() + '.csv\"');
-    const stringify = require('csv-stringify');
-    let ret = [[
-          `WikiRevId`,
-          `LastTimestamp`,
-          `OresDamagingScore`,
-          `OresBadfaithScore`,
-          `ShouldRevert`,
-          `NotSure`,
-          `LooksGood`
-        ]]
-        .concat(newJudgementCounts.map((newJudgementCount) => [
-          newJudgementCount.wikiRevId,
-          newJudgementCount.lastTimestamp,
-          newJudgementCount.recentChange.ores ? newJudgementCount.recentChange.ores.damagingScore: "null",
-          newJudgementCount.recentChange.ores ? newJudgementCount.recentChange.ores.badfaithScore: "null",
-          newJudgementCount.counts.ShouldRevert,
-          newJudgementCount.counts.NotSure,
-          newJudgementCount.counts.LooksGood,
-          // newJudgementCount.judgements
-        ]));
-    stringify(ret, {header: false})
-        .pipe(res);
-  }));
+  apiRouter.get("/markedRevs.csv", asyncHandler(routes.markedRevsCsv));
 
   /**
    * @deprecated
    */
-  apiRouter.get("/marked.csv", asyncHandler(async (req, res) => {
-    let myGaId = req.body.gaId || req.cookies._ga;
-    let recentChanges = await queryMarkedRecentChange(db, myGaId);
-    res.setHeader('Content-Type', 'text/csv');
-    res.setHeader('Content-Disposition', 'attachment; filename=\"' + 'download-' + Date.now() + '.csv\"');
-    // res.send(recentChanges);
-    const stringify = require('csv-stringify');
-    let ret = [[`WikiRevId`, `OresDamagingScore`, `OresBadfaithScore`, `ShouldRevert`, `NotSure`, `LooksGood`]]
-        .concat(recentChanges.map(rc => [
-          rc.wikiRevId, rc.ores.damagingScore, rc.ores.badfaithScore,
-          rc.judgementCounts.ShouldRevert || 0, rc.judgementCounts.NotSure || 0, rc.judgementCounts.LooksGood || 0
-        ]));
-    stringify(ret, {header: false})
-        .pipe(res);
-  }));
+  apiRouter.get("/marked.csv", asyncHandler(routes.markedCsv));
 
 
-  apiRouter.get("/markedRevs", asyncHandler(async (req, res) => {
-    res.send(await getNewJudgementCounts(db));
-  }));
+  apiRouter.get("/markedRevs", asyncHandler(routes.markedRevs));
 
   /**
    * @deprecated
    */
-  apiRouter.get("/marked", asyncHandler(async (req, res) => {
-    let myGaId = req.body.gaId || req.cookies._ga;
-    let recentChanges = await queryMarkedRecentChange(db, myGaId);
-    res.send(recentChanges);
-  }));
+  apiRouter.get("/marked", asyncHandler(routes.marked));
 
   /**
    * Return a list of all leader
@@ -784,422 +188,26 @@ function setupApiRequestListener(db, io, app) {
    *   SELECT user, count(*) FROM Interaction GROUP BY user ORDER by user;
    * ````
    */
-  apiRouter.get('/leaderboard', asyncHandler(async (req, res) => {
-    let myGaId = req.body.gaId || req.cookies._ga;
-
-    // TO-FUTURE-DO consider adding pagination if performance is a problem. We don't expect this list to be more than
-    // 10K records anytime soon
-    let ret = await db.collection("Interaction").aggregate(
-        [
-          {
-            "$match": {
-              "userGaId": {$ne: null}
-            }
-          },
-          {
-            "$group" : {
-              "_id" : {
-                "userGaId" : "$userGaId"
-              },
-              "count" : {
-                "$sum" : 1
-              },
-              "lastTimestamp" : {
-                "$max" : "$timestamp"
-              }
-            }
-          },
-          {
-            "$sort" : {
-              "count" : -1
-            }
-          },
-          {
-            "$project" : {
-              "userGaId" : "$_id.userGaId",
-              "count" : 1,
-              "lastTimestamp" : 1
-            }
-          }
-        ],
-        {
-          "allowDiskUse" : false
-        }
-    ).toArray();
-    res.send(ret);
-    req.visitor
-        .event({ec: "api", ea: "/leaderboard"})
-        .send();
-  }));
+  apiRouter.get('/leaderboard', asyncHandler(routes.leaderboard));
 
 
-  apiRouter.get('/stats', asyncHandler(async (req, res) => {
-    let myGaId = req.body.gaId || req.cookies._ga;
-    logger.debug(`req.query`, req.query);
-    let allInteractions = await db.collection(`Interaction`)
-        .find({}, {
-          userGaId: 1,
-          judgement: 1,
-          wikiRevId: 1,
-        }).toArray();
-    let revSet = {};
-    allInteractions.forEach(item => revSet[item.wikiRevId] = true);
-    let ret = {
-      totalJudgement: allInteractions.length,
-      totalRevJudged: Object.keys(revSet).length,
-      totalShouldRevert: allInteractions.filter(i => i.judgement === "ShouldRevert").length,
-    };
+  apiRouter.get('/stats', asyncHandler(routes.stats));
 
-    if (myGaId) {
-      let myInteractions = allInteractions.filter(i => i.userGaId === myGaId);
-      let myRevSet = {};
-      myInteractions.forEach(item => myRevSet[item.wikiRevId] = true);
-      ret.totalMyJudgement = myInteractions.length;
-      ret.totalMyRevJudged = Object.keys(myRevSet).length;
-      ret.totalMyShouldRevert = myInteractions.filter(item => item.judgement === "ShouldRevert").length;
-    }
+  // TODO build batch api for avatar until performance is an issue. We have cache anyway should be fine.
+  apiRouter.get("/avatar/:seed", asyncHandler(routes.avatar));
 
-    res.send(ret);
-    req.visitor
-        .event({ec: "api", ea: "/stats"})
-        .send();
-  }));
-
-  {
-    const Avatars = require('@dicebear/avatars').default;
-    const sprites = require('@dicebear/avatars-male-sprites').default;
-    let avatars = new Avatars(sprites({}));
-
-    // TODO build batch api for avatar until performance is an issue. We have cache anyway should be fine.
-    apiRouter.get("/avatar/:seed", asyncHandler(async (req, res) => {
-      logger.debug(`avatar requested with seed`, req.params.seed);
-      let svg = avatars.create(req.params.seed);
-      res.send(svg);
-      req.visitor
-          .event({ec: "api", ea: "/avatar/:seed"})
-          .send();
-    }));
-  }
-
-  /** Get the latest recentChangee
+  /** Get the latest recentChange
    * @deprecated
    */
-  apiRouter.get('/latest', asyncHandler(async (req, res) => {
+  apiRouter.get('/latest', asyncHandler(routes.latest));
 
-    let wiki = "enwiki";  // TODO: support multiple different wiki in the cases. Currently only support ENwiki.
+  apiRouter.get('/latestRevs', asyncHandler(routes.latestRevs));
 
-    // Getting a list of latest revisions related to the filter (Lang of Wiki), and their related diff
-    // TODO Consider use https://nodejs.org/api/url.html#url_url_searchparams to compose a standard one. this contains too many parameters
-    let queryUrl = `${req.query.serverUrl}/w/api.php?action=query&list=recentchanges&prop=info&format=json&rcnamespace=0&rclimit=100&rctype=edit&rctoponly=true&rcprop=user|userid|comment|flags|timestamp|ids|title&rcshow=!bot`;
-    // https://en.wikipedia.org/w/api.php?action=query&list=recentchanges&prop=info&format=json&rcnamespace=0&rclimit=50&rctype=edit&rctoponly=true&rcprop=user|userid|comment|flags|timestamp|ids|title&rcshow=!bot
-    let recentChangesJson = await rp.get(queryUrl, {json: true});
-    /** Sample response
-     {
-       "batchcomplete":"",
-       "continue":{
-          "rccontinue":"20190701214931|1167038199",
-          "continue":"-||info"
-       },
-       "query":{
-          "recentchanges":[
-             {
-                "type":"edit",
-                "ns":0,
-                "title":"Multiprocessor system architecture",
-                "pageid":58955273,
-                "revid":904396518,
-                "old_revid":904395753,
-                "rcid":1167038198,
-                "user":"Dhtwiki",
-                "userid":9475572,
-                "timestamp":"2019-07-01T21:49:32Z",
-                "comment":"Putting images at bottom, side by side, to prevent impinging on References section"
-             }
-             // ...
-          ]
-       }
-     }
+  apiRouter.get('/flags', routes.flags);
 
-     Converting to
-     {
-        _id: recentChange._id,
-        id: recentChange.id,
-        revision: recentChange.revision,
-        title: recentChange.title,
-        user: recentChange.user,
-        wiki: recentChange.wiki,
-        timestamp: recentChange.timestamp,
-        ores: recentChange.ores,
-        namespace: recentChange.namespace,
-        nonbot: !recentChange.bot
-      }
-     */
+  apiRouter.get('/mediawiki', asyncHandler(routes.mediawiki));
 
-    let oresUrl = `https://ores.wmflabs.org/v3/scores/${wiki}/?models=damaging|goodfaith&revids=${
-        recentChangesJson.query.recentchanges // from recentChangesJson result
-            .map(rawRecentChange => rawRecentChange.revid).join('|')
-        }`;
-
-    /**
-     {
-        enwiki:{
-          models:{
-            damaging:{
-              version:"0.5.0"
-            },
-            goodfaith:{
-              version:"0.5.0"
-            }
-          },
-          scores:{
-            904398217:{
-              damaging:{
-                score:{
-                  prediction:false,
-                  probability:{
-                    false:0.8201493218128969,
-                    true:0.1798506781871031
-                  }
-                }
-              },
-              goodfaith:{
-                score:{
-                  prediction:true,
-                  probability:{
-                    false:0.09864511980935009,
-                    true:0.9013548801906499
-                  }
-                }
-              }
-            },
-            904398221:{
-              damaging:{
-                score:{
-                  prediction:false,
-                  probability:{
-                    false:0.9593284228949122,
-                    true:0.04067157710508781
-                  }
-                }
-              },
-              goodfaith:{
-                score:{
-                  prediction:true,
-                  probability:{
-                    false:0.01102952942780866,
-                    true:0.9889704705721913
-                  }
-                }
-              }
-            }
-          }
-        }
-      }
-     * */
-
-    let oresResultJson = await rp.get(oresUrl, {json: true});
-
-    let recentChanges = recentChangesJson.query.recentchanges  // from recentChangesJson result
-        .map(rawRecentChange => {
-          return {
-            _id: `${wiki}-${rawRecentChange.rcid}`,
-            id: rawRecentChange.rcid,
-            revision: {
-              new: rawRecentChange.revid,
-              old: rawRecentChange.old_revid
-            },
-            title: rawRecentChange.title,
-            user: rawRecentChange.user,
-            ores: computeOresField(oresResultJson, wiki, rawRecentChange.revid),
-            wiki: `${wiki}`, // TODO verify
-            timestamp: Math.floor(new Date(rawRecentChange.timestamp).getTime() / 1000), // TODO check the exact format of timestamp. maybe use an interface?
-            namespace: 0, // we already query the server with "rcnamespace=0" filter
-            nonbot: true, // we already query the server with "rcprop=!bot" filter
-            comment: rawRecentChange.comment,
-          };
-        });
-    res.send(recentChanges.reverse());
-    req.visitor
-        .event({ec: "api", ea: "/latest"})
-        .send();
-
-  }));
-
-  apiRouter.get('/latestRevs', asyncHandler(async (req, res) => {
-    let wiki = "enwiki";  // TODO: support multiple different wiki in the cases. Currently only support ENwiki.
-
-    // Getting a list of latest revisions related to the filter (Lang of Wiki), and their related diff
-    // TODO Consider use https://nodejs.org/api/url.html#url_url_searchparams to compose a standard one. this contains too many parameters
-    let queryUrl = `${req.query.serverUrl}/w/api.php?action=query&list=recentchanges&prop=info&format=json&rcnamespace=0&rclimit=100&rctype=edit&rctoponly=true&rcprop=user|userid|comment|flags|timestamp|ids|title&rcshow=!bot`;
-    // https://en.wikipedia.org/w/api.php?action=query&list=recentchanges&prop=info&format=json&rcnamespace=0&rclimit=50&rctype=edit&rctoponly=true&rcprop=user|userid|comment|flags|timestamp|ids|title&rcshow=!bot
-    let recentChangesJson = await rp.get(queryUrl, {json: true});
-    /** Sample response
-     {
-       "batchcomplete":"",
-       "continue":{
-          "rccontinue":"20190701214931|1167038199",
-          "continue":"-||info"
-       },
-       "query":{
-          "recentchanges":[
-             {
-                "type":"edit",
-                "ns":0,
-                "title":"Multiprocessor system architecture",
-                "pageid":58955273,
-                "revid":904396518,
-                "old_revid":904395753,
-                "rcid":1167038198,
-                "user":"Dhtwiki",
-                "userid":9475572,
-                "timestamp":"2019-07-01T21:49:32Z",
-                "comment":"Putting images at bottom, side by side, to prevent impinging on References section"
-             }
-             // ...
-          ]
-       }
-     }
-
-     Converting to
-     {
-        _id: recentChange._id,
-        id: recentChange.id,
-        revision: recentChange.revision,
-        title: recentChange.title,
-        user: recentChange.user,
-        wiki: recentChange.wiki,
-        timestamp: recentChange.timestamp,
-        ores: recentChange.ores,
-        namespace: recentChange.namespace,
-        nonbot: !recentChange.bot
-      }
-     */
-
-    let oresUrl = `https://ores.wmflabs.org/v3/scores/${wiki}/?models=damaging|goodfaith&revids=${
-        recentChangesJson.query.recentchanges // from recentChangesJson result
-            .map(rawRecentChange => rawRecentChange.revid).join('|')
-        }`;
-
-    /**
-     {
-        enwiki:{
-          models:{
-            damaging:{
-              version:"0.5.0"
-            },
-            goodfaith:{
-              version:"0.5.0"
-            }
-          },
-          scores:{
-            904398217:{
-              damaging:{
-                score:{
-                  prediction:false,
-                  probability:{
-                    false:0.8201493218128969,
-                    true:0.1798506781871031
-                  }
-                }
-              },
-              goodfaith:{
-                score:{
-                  prediction:true,
-                  probability:{
-                    false:0.09864511980935009,
-                    true:0.9013548801906499
-                  }
-                }
-              }
-            },
-            904398221:{
-              damaging:{
-                score:{
-                  prediction:false,
-                  probability:{
-                    false:0.9593284228949122,
-                    true:0.04067157710508781
-                  }
-                }
-              },
-              goodfaith:{
-                score:{
-                  prediction:true,
-                  probability:{
-                    false:0.01102952942780866,
-                    true:0.9889704705721913
-                  }
-                }
-              }
-            }
-          }
-        }
-      }
-     * */
-
-    let oresResultJson = await rp.get(oresUrl, {json: true});
-
-    let recentChanges = recentChangesJson.query.recentchanges  // from recentChangesJson result
-        .map(rawRecentChange => {
-          return {
-            _id: `${wiki}-${rawRecentChange.rcid}`,
-            id: rawRecentChange.rcid,
-            wikiRevId: `${wiki}:${rawRecentChange.revid}`,
-            revision: {
-              new: rawRecentChange.revid,
-              old: rawRecentChange.old_revid
-            },
-            title: rawRecentChange.title,
-            user: rawRecentChange.user,
-            ores: computeOresField(oresResultJson, wiki, rawRecentChange.revid),
-            wiki: `${wiki}`, // TODO verify
-            timestamp: Math.floor(new Date(rawRecentChange.timestamp).getTime() / 1000), // TODO check the exact format of timestamp. maybe use an interface?
-            namespace: 0, // we already query the server with "rcnamespace=0" filter
-            nonbot: true, // we already query the server with "rcprop=!bot" filter
-            comment: rawRecentChange.comment,
-          };
-        });
-    res.send(recentChanges.reverse());
-    req.visitor
-        .event({ec: "api", ea: "/latestRevs"})
-        .send();
-
-  }));
-
-  apiRouter.get('/flags', (req, res, next) => {
-    res.send({
-      useStiki: useStiki,
-      useOauth: useOauth
-    });
-    req.visitor
-        .event({ec: "api", ea: "/"})
-        .send();
-  });
-
-  apiRouter.get('/mediawiki', asyncHandler(async (req, res) => {
-    // TODO add sanitize if we see abuse.
-    apiLogger.debug('req.params:', req.params);
-    apiLogger.debug('req.query:', req.query);
-    const fetchUrl = new URL(`${getUrlBaseByWiki(req.query.wiki)}/w/api.php`);
-    let params = JSON.parse(req.query.apiQuery);
-
-    Object.keys(params).forEach(key => {
-      fetchUrl.searchParams.set(key, params[key]);
-    });
-    let retJson = await rp.get(fetchUrl, {json: true});
-    res.send(retJson);
-    req.visitor
-        .event({ec: "mediawiki", ea: "/"})
-        .send();
-  }));
-
-  apiRouter.get('/version', (req, res, next) => {
-    var packageson = require('./../package.json');
-    res.send(packageson.version);
-    req.visitor
-        .event({ec: "api", ea: "/"})
-        .send();
-  });
+  apiRouter.get('/version', routes.version);
 
   app.use(`/api`, apiRouter);
 }
@@ -1411,9 +419,6 @@ async function start() {
   } else {
     await nuxt.ready()
   }
-  // Use connect method to connect to the Server
-  let db = (await MongoClient.connect(process.env.MONGODB_URI, {useNewUrlParser: true}))
-      .db(process.env.MONGODB_DB);
 
   app.use(function (req, res, next) {
     apiLogger.debug('req.originalUrl:', req.originalUrl);
